+++ src/app/(search)/search/page.tsx
@@ -1,10 +1,11 @@
 "use client";
 
-import { IMarkers, mapAtom, markerAtom } from "@/src/app/_atom/atom";
+import { mapAtom, markerAtom } from "@/src/app/_atom/atom";
 import { useRouter } from "next/navigation";
 import { useRecoilState, useSetRecoilState } from "recoil";
 import InfoCard from "../_component/InfoCard";
 import styles from "../_component/InfoCard.module.css";
+import { IMarkers } from "../../_interfaces/interface";
 
 function Home() {
   const router = useRouter();
