+++ src/app/_component/SearchNav/SearchNav.tsx
@@ -7,11 +7,12 @@ import BusIcon from "@/public/images/bus.svg";
 import SubwayIcon from "@/public/images/subway.svg";
 import SearchIcon from "@/public/images/search.svg";
 import { useRouter, useSelectedLayoutSegments } from "next/navigation";
-import { IMarkers, mapAtom, pageState } from "../../_atom/atom";
+import { mapAtom, pageState } from "../../_atom/atom";
 import { useRecoilState } from "recoil";
 import cx from "classnames";
 import { useEffect, useState } from "react";
 import { judgeActivation } from "@/src/app/_utils/navigationUtils";
+import { IMarkers } from "../../_interfaces/interface";
 
 const inter = Inter({ subsets: ["latin"] });
